refactor: renamed/moved files for consistency
